% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\title{Finds correlation matrix of all cols for inputted year.}
\usage{
corr_matrix(df, yr)
}
\arguments{
\item{df}{A dataframe.}

\item{yr}{A year.}
}
\value{
correlation matrix.
}
\description{
Finds correlation matrix of all cols for inputted year.
}
\examples{
corr_matrix(nbastats, 1993)
}
