% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_age.R
\name{max_age}
\alias{max_age}
\title{Finds max age of player provided inputted year.}
\usage{
max_age(df, yr)
}
\arguments{
\item{df}{A dataframe.}

\item{yr}{A year.}
}
\value{
top player.
}
\description{
Finds max age of player provided inputted year.
}
\examples{
max_age(nbastats, 1993)
add(10, 1)
}
