% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_points.R
\name{max_points}
\alias{max_points}
\title{Finds max points and player provided inputted year.}
\usage{
max_points(df, yr)
}
\arguments{
\item{df}{A dataframe.}

\item{yr}{A year.}
}
\value{
top player and max points.
}
\description{
Finds max points and player provided inputted year.
}
\examples{
max_points(nbastats, 1993)
}
